requirements.txt
fastapi==0.115.4
google-generativeai==0.8.3
pydantic==2.9.2
mangum==0.18.0
